cleaned up code and added some config logic

+++ pyrad/engine/trainer.py
@@ -22,10 +22,11 @@ from typing import Dict
 
 import torch
 import torch.distributed as dist
+import umsgpack
 from torch.nn.parallel import DistributedDataParallel as DDP
 from torchtyping import TensorType
-from pyrad.cameras.cameras import get_camera, get_intrinsics_from_intrinsics_matrix
 
+from pyrad.cameras.cameras import get_camera, get_intrinsics_from_intrinsics_matrix
 from pyrad.cameras.rays import RayBundle
 from pyrad.data.dataloader import EvalDataloader, setup_dataset_eval, setup_dataset_train
 from pyrad.graphs.base import setup_graph
@@ -34,11 +35,9 @@ from pyrad.utils import profiler, writer
 from pyrad.utils.config import Config
 from pyrad.utils.decorators import check_main_thread
 from pyrad.utils.writer import EventName, TimeWriter
-from pyrad.viewer.backend.visualizer import ViewerWindow, Visualizer
-from pyrad.viewer.backend.utils import get_intrinsics_matrix_and_camera_to_world_h
 from pyrad.viewer.backend import vis_utils
-
-import umsgpack
+from pyrad.viewer.backend.utils import get_intrinsics_matrix_and_camera_to_world_h
+from pyrad.viewer.backend.visualizer import ViewerWindow, Visualizer
 
 logging.getLogger("PIL").setLevel(logging.WARNING)
 
@@ -47,10 +46,9 @@ class Trainer:
     """Training class
 
     Args:
-        config (Config): _description_
-        local_rank (int, optional): _description_. Defaults to 0.
-        world_size (int, optional): _description_. Defaults to 1.
-        cpu (bool, optional): Whether or not to use the CPU.
+        config (Config): The configuration object.
+        local_rank (int, optional): Local rank of the process. Defaults to 0.
+        world_size (int, optional): World size of the process. Defaults to 1.
     """
 
     def __init__(self, config: Config, local_rank: int = 0, world_size: int = 1):
@@ -70,10 +68,14 @@ class Trainer:
         writer.setup_event_writers(config.logging, max_iter=config.trainer.max_num_iterations)
         profiler.setup_profiler(config.logging)
 
-        window = ViewerWindow(zmq_url=self.config.viewer.zmq_url)
-        self.vis = Visualizer(window=window)
-        logging.info("Connected to viewer")
-        self.vis.delete()
+        self.vis = None
+        if self.config.viewer.enable:
+            window = ViewerWindow(zmq_url=self.config.viewer.zmq_url)
+            self.vis = Visualizer(window=window)
+            logging.info("Connected to viewer at %s", self.config.viewer.zmq_url)
+            self.vis.delete()
+        else:
+            logging.info("Continuing without viewer.")
 
     def setup(self, test_mode=False):
         """Setup the Trainer by calling other setup functions."""
@@ -105,7 +107,8 @@ class Trainer:
     def train(self) -> None:
         """Train the model."""
 
-        self.draw_scene_in_viewer()
+        if self.vis:
+            self.draw_scene_in_viewer()
 
         with TimeWriter(writer, EventName.TOTAL_TRAIN_TIME):
             num_iterations = self.config.trainer.max_num_iterations
@@ -123,7 +126,8 @@ class Trainer:
                 if step != 0 and self.config.trainer.steps_per_save and step % self.config.trainer.steps_per_save == 0:
                     self._save_checkpoint(self.config.trainer.model_dir, step)
                 if (
-                    step != 0
+                    self.vis
+                    and step != 0
                     and self.config.viewer.steps_per_render_image
                     and step % self.config.viewer.steps_per_render_image == 0
                 ):
@@ -238,7 +242,7 @@ class Trainer:
                 pose=pose,
                 K=K,
                 height=1.0,
-                name="image_dataset_train/{:06d}".format(idx),
+                name=f"image_dataset_train/{idx:06d}",
                 displayed_focal_length=0.5,
                 realistic=False,
             )
@@ -246,14 +250,15 @@ class Trainer:
         vis_utils.draw_aabb(self.vis, aabb, name="dataset_inputs_train/scene_bounds/aabb")
 
     @profiler.time_function
-    def render_image_in_viewer(self, num_rays_per_chunk=4096):
-        # TODO: remove the hardcoded num_rays_per_chunk
-        # print("render_image_in_viewer")
+    def render_image_in_viewer(self):
+        """
+        Draw an image using the current camera pose from the viewer.
+        The image is sent of a TCP connection and then uses WebRTC to send it to the viewer.
+        """
         data = self.vis["/Cameras/Main Camera"].get_object()
         message = umsgpack.unpackb(data)
         camera_object = message["object"]["object"]
-        # TODO: remove the notion of the hardcoded image_height below
-        image_height = 800
+        image_height = self.config.viewer.render_image_height
         intrinsics_matrix, camera_to_world_h = get_intrinsics_matrix_and_camera_to_world_h(
             camera_object, image_height=image_height
         )
@@ -262,16 +267,16 @@ class Trainer:
         intrinsics = get_intrinsics_from_intrinsics_matrix(intrinsics_matrix)
         camera = get_camera(intrinsics, camera_to_world)
         camera_ray_bundle = camera.get_camera_ray_bundle(device=self.device)
-        camera_ray_bundle.num_rays_per_chunk = num_rays_per_chunk
+        camera_ray_bundle.num_rays_per_chunk = self.config.viewer.num_rays_per_chunk
 
         self.graph.eval()
         outputs = self.graph.get_outputs_for_camera_ray_bundle(camera_ray_bundle)
         self.graph.train()
 
-        # gross hack to get the image key
+        # gross hack to get the image key, depending on which keys the graph uses
         rgb_key = "rgb" if "rgb" in outputs else "rgb_fine"
+        # TODO: make it such that the TCP connection doesn't need float64
         image = outputs[rgb_key].cpu().numpy().astype("float64") * 255
-        print(image.shape)
         self.vis["/Cameras/Main Camera"].set_image(image)
         return outputs
